Zend compatible cachine changes: Modified HttpRequest to use a caching object compatible with Zend_Cache HttpRequestFactory creates a default file based cache according to usual moriarty constant definitions but this can be overridden to create a specalisation, e.g. for memcached Each domain object now takes an optional HttpRequestFactory as a parameter on its constructor

+++ trunk/httpcache.class.php
@@ -3,73 +3,77 @@ require_once dirname(__FILE__) . DIRECTORY_SEPARATOR . 'moriarty.inc.php';
 
 class HttpCache {
   var $_directory;
+  var $_options;
   
-  function __construct($directory) {
-    $this->_directory = $directory;
+  function __construct($options = array()) {
+    $this->_options = $options;
+    $this->_directory = $options['directory'];
     if (substr($this->_directory, -1) != DIRECTORY_SEPARATOR) {
       $this->_directory .= DIRECTORY_SEPARATOR;
     }
+    
   }
 
-
-  function make_conditional_request($request) {
-    return $request;
-  }
-  
-  function write($request, $response) {
-    $filename = $this->get_cache_filename($request);
+ 
+  function save($data, $id, $tags= array(), $specificLifetime = FALSE, $priority = 0) {
+    if ($this->_options['caching'] === FALSE) return TRUE;
+    $filename = $this->_directory . $id;
     $fp = fopen($filename, 'w');
     if ($fp) {
-      fwrite($fp,serialize($response));
+      fwrite($fp,serialize($specificLifetime) . "\n");
+      fwrite($fp,serialize($data));
       fclose($fp);    
       chmod($filename, 0777);
+      return TRUE;
     }
+    return FALSE;
   }
 
-  function get_cached_response($request) {
-    $filename = $this->get_cache_filename($request);
+  function load($id, $doNotTestCacheValidity = FALSE, $doNotUnserialize = FALSE) {
+    if ($this->_options['caching'] === FALSE) return FALSE;
+    $filename = $this->_directory . $id;
     if ( file_exists($filename)) {
       $content = file_get_contents($filename);
       if ($content !== FALSE ) {
-        $response = unserialize($content);
+        list($specificLifetime_ser , $response_ser) = explode("\n", $content, 2);
+        $specificLifetime= unserialize($specificLifetime_ser);
+        if (! $doNotTestCacheValidity) {
+          if ( time() - filectime($filename) > $specificLifetime ) {
+            $this->remove($id);
+            return FALSE;
+          } 
+        }
+        $response = unserialize($response_ser);
         return $response;
       }       
     }
-    return null;
+    return FALSE;
   }
-  
-  function is_fresh(&$request, &$response) {
-    $filename = $this->get_cache_filename($request);
-    $cache_control = $response->headers['cache-control'];
-    $cache_control_tokens = split(',', $cache_control);
-    foreach ( $cache_control_tokens as $token) {
-      $token = trim($token);
-      if ( preg_match('/max-age=(.+)/', $token, $m) ) {
-        $max_age = $m[1];
-        if ( time() - filectime($filename) > $max_age ) {
-          return false;
-        } 
-      }
-    }
-    
-    return true;
-  }
-  
-  function remove_from_cache($request) {
-    $filename = $this->get_cache_filename($request);
+
+  function remove($id) {
+    if ($this->_options['caching'] === FALSE) return TRUE;
+
+    $filename = $this->_directory . $id;
     if ( file_exists($filename)) {
       unlink($filename);
-    }   
+    }       
+
+    return TRUE;
+  }
+
+
+  function test($id) {
+    if ($this->_options['caching'] === FALSE) return FALSE;
+    $filename = $this->_directory . $id;  
+    return file_exists($filename);
   }
   
-  function get_cache_filename($request) {
-    $accept = $request->headers['Accept'];
-    $accept_parts = split(',', $accept);
-    sort($accept_parts);
-    $accept = join(',', $accept_parts);
-    return $this->_directory . md5('<' . $request->uri . '>' . $accept . $request->get_body()); 
-  }
   
 
 }
+
+
+
+
+
 ?>
